requirements.txt
flask==2.0.1
gunicorn==20.1.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.28.0
google-cloud-storage==1.42.3